feat: error treatments inplamentation

+++ processa_exclusao_produto.php
@@ -5,16 +5,27 @@ include 'conecta.php';
 if (isset($_POST['id_produto'])) {
     $id_produto = $_POST['id_produto'];
 
-    // Remove o produto da tabela
-    $sql = "DELETE FROM produtos WHERE id_produto = $id_produto";
+    // Consulta para verificar se o produto existe no banco de dados
+    $consulta_produto = "SELECT * FROM produtos WHERE id_produto = $id_produto";
+    $resultado_consulta = mysqli_query($sistemas_vendas, $consulta_produto);
 
-    if (mysqli_query($sistemas_vendas, $sql)) {
-        echo "Produto excluído com sucesso!";
-        // Redirecionamento após a exclusão
-        header("Location: index.php");
-        exit();
+    if (mysqli_num_rows($resultado_consulta) > 0) {
+        // Produto encontrado, prosseguir com a exclusão
+        $sql = "DELETE FROM produtos WHERE id_produto = $id_produto";
+
+        if (mysqli_query($sistemas_vendas, $sql)) {
+            echo "Produto excluído com sucesso!";
+            // Redirecionamento após a exclusão
+            header("refresh:3; url=index.php");
+            exit();
+        } else {
+            echo "Erro ao excluir produto: " . mysqli_error($sistemas_vendas);
+        }
     } else {
-        echo "Erro ao excluir produto: " . mysqli_error($sistemas_vendas);
+        // Produto não encontrado, informar ao usuário e redirecionar para a página de listagem
+        echo "Produto não encontrado.";
+        header("refresh:3; url=index.php");
+        exit();
     }
 
     // Fecha a conexão com o banco de dados
